Move MetadataMixin to InvenTree.models
To prevent circular imports

(cherry picked from commit d23b013881eaffe612dfbfcdfc5dff6d729068c6)
