+++ InvenTree/company/models.py
@@ -26,9 +26,8 @@ import InvenTree.validators
 from common.settings import currency_code_default
 from InvenTree.fields import InvenTreeURLField, RoundingDecimalField
 from InvenTree.models import (InvenTreeAttachment, InvenTreeBarcodeMixin,
-                              InvenTreeNotesMixin)
+                              InvenTreeNotesMixin, MetadataMixin)
 from InvenTree.status_codes import PurchaseOrderStatus
-from plugin.models import MetadataMixin
 
 
 def rename_company_image(instance, filename):
