+++ InvenTree/label/models.py
@@ -17,7 +17,7 @@ import common.models
 import part.models
 import stock.models
 from InvenTree.helpers import normalize, validateFilterString
-from plugin.models import MetadataMixin
+from InvenTree.models import MetadataMixin
 
 try:
     from django_weasyprint import WeasyTemplateResponseMixin
